+++ project_book/src/main/java/com/example/project_book/model/Users.java
@@ -1,4 +1,4 @@
-package com.example.bookstore.model;
+package com.example.project_book.model;
 
 import javax.persistence.*;
 import javax.validation.constraints.Email;
