+++ project_book/src/main/java/com/example/project_book/repo/IUsersTypeRepo.java
@@ -1,6 +1,6 @@
-package com.example.bookstore.repo;
+package com.example.project_book.repo;
 
-import com.example.bookstore.model.UsersType;
+import com.example.project_book.model.UsersType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface IUsersTypeRepo extends JpaRepository<UsersType,Integer> {
