+++ project_book/src/main/java/com/example/project_book/service/IUsersService.java
@@ -1,6 +1,6 @@
-package com.example.bookstore.service;
+package com.example.project_book.service;
 
-import com.example.bookstore.model.Users;
+import com.example.project_book.model.Users;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
